x86: x86 CSV table and decoder generator
This CL contains the x86 CSV table and the decoder generator that reads it.
The program to generate the CSV table is not included (yet),
as explained in the comments at the top of the table.

The generator is used to generate the decoder in the x86/x86asm package
(separate CL).

LGTM=crawshaw
R=crawshaw
https://codereview.appspot.com/91170044